refactor: exception, dto 패키지 분리
