+++ src/main/java/kr/codesqaud/cafe/controller/dto/article/ArticleInfoDTO.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.controller.dto;
+package kr.codesqaud.cafe.controller.dto.article;
 
 public class ArticleInfoDTO {
