+++ src/main/java/kr/codesqaud/cafe/controller/dto/article/ArticleUpdateDTO.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.controller.dto;
+package kr.codesqaud.cafe.controller.dto.article;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Pattern;
