+++ src/main/java/kr/codesqaud/cafe/controller/dto/user/UserListDTO.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.controller.dto;
+package kr.codesqaud.cafe.controller.dto.user;
 
 public class UserListDTO {
     private final String nickName;
