+++ src/main/java/kr/codesqaud/cafe/exception/article/ArticleIdAndSessionIdMismatchException.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.exception;
+package kr.codesqaud.cafe.exception.article;
 
 public class ArticleIdAndSessionIdMismatchException extends RuntimeException{
     private static final String ARTICLE_ID_AND_SESSION_ID_MISMATCH_EXCEPTION = "수정또는 삭제 할 수 있는 권한이 없습니다.";
