+++ src/main/java/kr/codesqaud/cafe/exception/user/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.exception;
+package kr.codesqaud.cafe.exception.user;
 
 public class UserNotFoundException extends RuntimeException{
