+++ src/main/java/kr/codesqaud/cafe/exception/user/UserUpdateInvalidPasswordException.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.exception;
+package kr.codesqaud.cafe.exception.user;
 
 public class UserUpdateInvalidPasswordException extends RuntimeException{
     private static final String USER_UPDATE_INVALID_PASSWORD_EXCEPTION = "비밀번호가 일치하지 않습니다.";
